# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus Sentinel API",
    "description": "UEBA sobre datos reales de LANL. Sirve la cola de triage priorizada por riesgo (0-100), siempre acompañada de su explicación SHAP. Ninguna acción sobre cuentas es automática: el sistema prioriza y explica, la decisión es humana.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/employees": {
      "get": {
        "tags": [
          "triage"
        ],
        "summary": "Employees",
        "description": "Cola de triage priorizada (usuarios anonimizados U-####).",
        "operationId": "employees_employees_get",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(risk|alertas)$",
              "default": "risk",
              "title": "Sort"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Nivel"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por nivel del semáforo",
              "title": "Level"
            },
            "description": "Filtra por nivel del semáforo"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmpleadoCola"
                  },
                  "title": "Response Employees Employees Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{employee_id}/risk": {
      "get": {
        "tags": [
          "empleado"
        ],
        "summary": "Employee Risk",
        "description": "Serie temporal de la puntuación de riesgo (últimos `days` días disponibles).",
        "operationId": "employee_risk_employees__employee_id__risk_get",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerieRiesgo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{employee_id}/explanation": {
      "get": {
        "tags": [
          "empleado"
        ],
        "summary": "Employee Explanation",
        "description": "Top-5 SHAP del día, con el valor del usuario vs. su promedio personal y vs. su peer group.\n\nSolo hay explicación para días con alerta: los verdes son registro pasivo.",
        "operationId": "employee_explanation_employees__employee_id__explanation_get",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Explicacion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{employee_id}/activity": {
      "get": {
        "tags": [
          "empleado"
        ],
        "summary": "Employee Activity",
        "description": "Resumen de autenticaciones del día: destinos nuevos, fallos, NTLM y horario.",
        "operationId": "employee_activity_employees__employee_id__activity_get",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actividad"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/feedback": {
      "post": {
        "tags": [
          "triage"
        ],
        "summary": "Case Feedback",
        "description": "Registra la decisión del analista (falso positivo / investigar / escalar).\n\nCierra el ciclo human-in-the-loop: ninguna acción sobre la cuenta es automática, y la decisión\nqueda trazada para auditoría.\nNota de despliegue: en planes gratuitos el almacenamiento es efímero.",
        "operationId": "case_feedback_cases__case_id__feedback_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "triage"
        ],
        "summary": "Case Feedback History",
        "description": "Historial de decisiones del analista sobre un caso.\n\nCierra la trazabilidad que exige un perfil de cumplimiento: toda decisión sobre\nuna cuenta queda registrada y es consultable.",
        "operationId": "case_feedback_history_cases__case_id__feedback_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackRegistro"
                  },
                  "title": "Response Case Feedback History Cases  Case Id  Feedback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/assistant/chat": {
      "post": {
        "tags": [
          "asistente"
        ],
        "summary": "Assistant Chat",
        "description": "Asistente conversacional: guía al usuario y responde preguntas sobre los datos\nconsultando el modelo mediante function calling (DeepSeek).\n\nLa API key vive solo en el backend. Sin ella, degrada con 503 explícito.",
        "operationId": "assistant_chat_assistant_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inference/score": {
      "post": {
        "tags": [
          "inferencia"
        ],
        "summary": "Inference Score",
        "description": "Inferencia REAL sobre un usuario-día: ejecuta el modelo en vivo.\n\nDemuestra el flujo completo frente al snapshot precomputado. Requiere los artefactos\nserializados; si no están presentes responde 503 en lugar de fallar de forma opaca.",
        "operationId": "inference_score_inference_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenciaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenciaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Actividad": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "n_eventos": {
            "type": "integer",
            "title": "N Eventos"
          },
          "n_dst": {
            "type": "integer",
            "title": "N Dst"
          },
          "n_destinos_nuevos": {
            "type": "integer",
            "title": "N Destinos Nuevos"
          },
          "n_origenes_nuevos": {
            "type": "integer",
            "title": "N Origenes Nuevos"
          },
          "n_fallos": {
            "type": "integer",
            "title": "N Fallos"
          },
          "ratio_ntlm": {
            "type": "number",
            "title": "Ratio Ntlm"
          },
          "n_fuera_horario": {
            "type": "integer",
            "title": "N Fuera Horario"
          },
          "ego_graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EgoGraph"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "day",
          "n_eventos",
          "n_dst",
          "n_destinos_nuevos",
          "n_origenes_nuevos",
          "n_fallos",
          "ratio_ntlm",
          "n_fuera_horario"
        ],
        "title": "Actividad"
      },
...
      "EmpleadoCola": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID anonimizado",
            "examples": [
              "U-0737"
            ]
          },
          "risk_max": {
            "type": "number",
            "title": "Risk Max"
          },
          "level": {
            "$ref": "#/components/schemas/Nivel"
          },
          "dia_pico": {
            "type": "integer",
            "title": "Dia Pico"
          },
          "n_dias_alerta": {
            "type": "integer",
            "title": "N Dias Alerta"
          },
          "motivo": {
            "type": "string",
            "title": "Motivo",
            "description": "Razón principal en una línea, derivada del top SHAP"
          },
          "es_comprometida": {
            "type": "integer",
            "title": "Es Comprometida",
            "description": "Ground truth (demo académica)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "risk_max",
          "level",
          "dia_pico",
          "n_dias_alerta",
          "motivo",
          "es_comprometida"
        ],
        "title": "EmpleadoCola"
      },
      "Explicacion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "risk": {
            "type": "number",
            "title": "Risk"
          },
          "level": {
            "$ref": "#/components/schemas/Nivel"
          },
          "prob": {
            "type": "number",
            "title": "Prob"
          },
          "motivo": {
            "type": "string",
            "title": "Motivo"
          },
          "top_shap": {
            "items": {
              "$ref": "#/components/schemas/ContribucionShap"
            },
            "type": "array",
            "title": "Top Shap"
          }
        },
        "type": "object",
        "required": [
          "id",
          "day",
          "risk",
          "level",
          "prob",
          "motivo",
          "top_shap"
        ],
        "title": "Explicacion"
      },
      "FeedbackIn": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "analista": {
            "type": "string",
            "maxLength": 64,
            "title": "Analista",
            "default": "anonimo"
          },
          "dia": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dia"
          },
          "nota": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Nota"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "FeedbackIn"
      },
      "FeedbackOut": {
        "properties": {
          "registrado": {
            "type": "boolean",
            "title": "Registrado"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "registrado",
          "case_id",
          "decision",
          "timestamp"
        ],
        "title": "FeedbackOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Nivel": {
        "type": "string",
        "enum": [
          "verde",
          "naranja",
          "rojo"
        ],
        "title": "Nivel"
      },
...
      "SerieRiesgo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "peer_cluster": {
            "type": "integer",
            "title": "Peer Cluster"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/PuntoRiesgo"
            },
            "type": "array",
            "title": "Timeline"
          }
        },
        "type": "object",
        "required": [
          "id",
          "peer_cluster",
          "timeline"
        ],
        "title": "SerieRiesgo"
      },
...
    }
  }
}